create_db: switch to float type

+++ create_db.sql
@@ -11,9 +11,9 @@ USE logmasimo;
 CREATE TABLE `data` (
 	`id` int(11) NOT NULL AUTO_INCREMENT COMMENT 'Unique Identifier',
 	`ts` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'Time stamp of data',
-	`spo2` int(11) DEFAULT 0 COMMENT 'O2 % (peripheral capillary oxygen saturation)',
-	`bpm` int(11) DEFAULT 0 COMMENT 'Beats Per Minute',
-	`pi` int(11) DEFAULT 0 COMMENT 'Perfusion Index',
+	`spo2` int(3) DEFAULT 0 COMMENT 'O2 % (peripheral capillary oxygen saturation)',
+	`bpm` int(5) DEFAULT 0 COMMENT 'Beats Per Minute',
+	`pi` float DEFAULT 0 COMMENT 'Perfusion Index',
 	`alarm` int(11) DEFAULT 0 COMMENT 'Raw Alarm value',
 	`exc` int(11) DEFAULT 0 COMMENT 'Raw Exception',
 	`exc1` int(11) DEFAULT 0 COMMENT 'Raw Exception 1',
